requirement.txt
numpy
pandas
matplotlib
plotly
scikit-learn
scipy
yfinance
pydantic
gnews
openai
instructor
redis
python-dotenv
